requirements.txt
mcp>=1.0.0
websockets>=12.0
python-dotenv>=1.0.1
pydantic>=2.7.0
